fix(canvas): wheel + space-pan drive container scroll; split oversize files
Master-QA on the prior #66 commit flagged three blockers against the
issue's expected behaviour. Resolved here:

1. Mouse wheel / shift+wheel / trackpad two-finger scroll didn't scroll
   the canvas viewport. The Fabric `mouse:wheel` handler called
   `preventDefault` on every wheel event and translated `viewportTransform`,
   which under the new `page * zoom` canvas-sizing model shifted Fabric's
   drawing inside its own pixel area instead of moving the visible
   viewport. `wireWheelEvents` now early-returns for non-zoom wheels so
   the browser's native scroll on the `overflow: auto` container handles
   them. Ctrl/meta+wheel still zooms.

2. Space+drag pan was broken by the same model change — translating the
   transform was wiped on the next zoom-sync. `wireMouseEvents` pan now
   drives `containerRef.current.scrollLeft/Top` instead, so it tracks
   the scrollbar position and survives every effect re-run.

3. Hard Rule #11 (300 LOC cap) was violated on both touched files.
   Split `useFabricCanvas.ts` (439 → 140) into the hook plus
   `wireSelectionEvents`, `wireDragResizeEvents`, `wireMouseEvents`,
   `wireWheelEvents`. Split `CanvasArea.tsx` (311 → 248) by extracting
   `useCurrentBackground` to its own module.

Tests: three new e2e — vertical wheel scroll, shift+wheel horizontal
scroll, and space+drag scrolling the container — that would have caught
blockers 1 and 2.
